backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
Werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
redis==5.0.1
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
bcrypt==4.1.2
# psycopg2-binary==2.9.9  # Temporarily removed for Windows development
alembic==1.13.1
stripe==12.4.0
Flask-Mail==0.9.1